requirements.txt
pyfiglet==0.8.post1
termcolor==1.1.0
pycryptodome==3.9.7
